Silero_pipeline/requirements.txt
torch
pyyaml
soundfile
librosa
tqdm
pandas
scikit-learn
# Библиотека для работы с git-репозиториями из Python
GitPython